make record/encode working on iOS
- use atomic bitset/bitdrop
- drop resampler from iOS code path
- use time from double
- add "goto restart" back to sender code

+++ libresip/ajitter.h
@@ -8,7 +8,7 @@ typedef struct
 	int used;
 	char *buffer;
 	double time[AJD];
-	int last;
+	double last;
 	int csize;
 	char *out_buffer;
 	int out_have;
